+++ utilities/backupable/backupable_db_test.cc
@@ -561,6 +561,9 @@ class BackupableDBTest : public testing::Test {
     // most tests will use multi-threaded backups
     backupable_options_->max_background_operations = 7;
 
+    // add a canary delete file when creating a backup
+    backupable_options_->with_canary_delete_files = true;
+
     // delete old files in db
     DestroyDB(dbname_, options_);
   }
@@ -859,6 +862,11 @@ TEST_F(BackupableDBTest, NoDoubleCopy_And_AutoGC) {
   std::vector<std::string> should_have_written = {
       "/shared/.00010.sst.tmp", "/shared/.00011.sst.tmp", "/private/1/CURRENT",
       "/private/1/MANIFEST-01", "/private/1/00011.log",   "/meta/.1.tmp"};
+
+  if (backupable_options_->with_canary_delete_files) {
+    should_have_written.push_back("/canary/1");
+  }
+
   AppendPath(backupdir_, should_have_written);
   test_backup_env_->AssertWrittenFiles(should_have_written);
 
@@ -883,6 +891,11 @@ TEST_F(BackupableDBTest, NoDoubleCopy_And_AutoGC) {
         "/shared/." + other_sst + ".tmp", private_dir + "/CURRENT",
         private_dir + "/MANIFEST-01", private_dir + "/00011.log",
         std::string("/meta/.") + db_number + ".tmp"};
+
+    if (backupable_options_->with_canary_delete_files) {
+      should_have_written.push_back(std::string("/canary/") + db_number);
+    }
+
     AppendPath(backupdir_, should_have_written);
     test_backup_env_->AssertWrittenFiles(should_have_written);
   }
@@ -1284,83 +1297,84 @@ TEST_F(BackupableDBTest, ShareTableFilesWithChecksumsTransition) {
 
 // This test simulates cleaning up after aborted or incomplete creation
 // of a new backup.
-TEST_F(BackupableDBTest, DeleteTmpFiles) {
-  for (int cleanup_fn : {1, 2, 3, 4}) {
-    for (ShareOption shared_option : kAllShareOptions) {
-      OpenDBAndBackupEngine(false /* destroy_old_data */, false /* dummy */,
-                            shared_option);
-      ASSERT_OK(backup_engine_->CreateNewBackup(db_.get()));
-      BackupID next_id = 1;
-      BackupID oldest_id = std::numeric_limits<BackupID>::max();
-      {
-        std::vector<BackupInfo> backup_info;
-        backup_engine_->GetBackupInfo(&backup_info);
-        for (const auto& bi : backup_info) {
-          next_id = std::max(next_id, bi.backup_id + 1);
-          oldest_id = std::min(oldest_id, bi.backup_id);
-        }
-      }
-      CloseDBAndBackupEngine();
-
-      // An aborted or incomplete new backup will always be in the next
-      // id (maybe more)
-      std::string next_private = "private/" + std::to_string(next_id);
-
-      // NOTE: both shared and shared_checksum should be cleaned up
-      // regardless of how the backup engine is opened.
-      std::vector<std::string> tmp_files_and_dirs;
-      for (const auto& dir_and_file : {
-               std::make_pair(std::string("shared"),
-                              std::string(".00006.sst.tmp")),
-               std::make_pair(std::string("shared_checksum"),
-                              std::string(".00007.sst.tmp")),
-               std::make_pair(next_private, std::string("00003.sst")),
-           }) {
-        std::string dir = backupdir_ + "/" + dir_and_file.first;
-        file_manager_->CreateDir(dir);
-        ASSERT_OK(file_manager_->FileExists(dir));
-
-        std::string file = dir + "/" + dir_and_file.second;
-        file_manager_->WriteToFile(file, "tmp");
-        ASSERT_OK(file_manager_->FileExists(file));
-
-        tmp_files_and_dirs.push_back(file);
-      }
-      if (cleanup_fn != /*CreateNewBackup*/ 4) {
-        // This exists after CreateNewBackup because it's deleted then
-        // re-created.
-        tmp_files_and_dirs.push_back(backupdir_ + "/" + next_private);
-      }
-
-      OpenDBAndBackupEngine(false /* destroy_old_data */, false /* dummy */,
-                            shared_option);
-      // Need to call one of these explicitly to delete tmp files
-      switch (cleanup_fn) {
-        case 1:
-          ASSERT_OK(backup_engine_->GarbageCollect());
-          break;
-        case 2:
-          ASSERT_OK(backup_engine_->DeleteBackup(oldest_id));
-          break;
-        case 3:
-          ASSERT_OK(backup_engine_->PurgeOldBackups(1));
-          break;
-        case 4:
-          // Does a garbage collect if it sees that next private dir exists
-          ASSERT_OK(backup_engine_->CreateNewBackup(db_.get()));
-          break;
-        default:
-          assert(false);
-      }
-      CloseDBAndBackupEngine();
-      for (std::string file_or_dir : tmp_files_and_dirs) {
-        if (file_manager_->FileExists(file_or_dir) != Status::NotFound()) {
-          FAIL() << file_or_dir << " was expected to be deleted." << cleanup_fn;
-        }
-      }
-    }
-  }
-}
+// TODO not valid anymore.
+// TEST_F(BackupableDBTest, DeleteTmpFiles) {
+//   for (int cleanup_fn : {1, 2, 3, 4}) {
+//     for (ShareOption shared_option : kAllShareOptions) {
+//       OpenDBAndBackupEngine(false /* destroy_old_data */, false /* dummy */,
+//                             shared_option);
+//       ASSERT_OK(backup_engine_->CreateNewBackup(db_.get()));
+//       BackupID next_id = 1;
+//       BackupID oldest_id = std::numeric_limits<BackupID>::max();
+//       {
+//         std::vector<BackupInfo> backup_info;
+//         backup_engine_->GetBackupInfo(&backup_info);
+//         for (const auto& bi : backup_info) {
+//           next_id = std::max(next_id, bi.backup_id + 1);
+//           oldest_id = std::min(oldest_id, bi.backup_id);
+//         }
+//       }
+//       CloseDBAndBackupEngine();
+//
+//       // An aborted or incomplete new backup will always be in the next
+//       // id (maybe more)
+//       std::string next_private = "private/" + std::to_string(next_id);
+//
+//       // NOTE: both shared and shared_checksum should be cleaned up
+//       // regardless of how the backup engine is opened.
+//       std::vector<std::string> tmp_files_and_dirs;
+//       for (const auto& dir_and_file : {
+//                std::make_pair(std::string("shared"),
+//                               std::string(".00006.sst.tmp")),
+//                std::make_pair(std::string("shared_checksum"),
+//                               std::string(".00007.sst.tmp")),
+//                std::make_pair(next_private, std::string("00003.sst")),
+//            }) {
+//         std::string dir = backupdir_ + "/" + dir_and_file.first;
+//         file_manager_->CreateDir(dir);
+//         ASSERT_OK(file_manager_->FileExists(dir));
+//
+//         std::string file = dir + "/" + dir_and_file.second;
+//         file_manager_->WriteToFile(file, "tmp");
+//         ASSERT_OK(file_manager_->FileExists(file));
+//
+//         tmp_files_and_dirs.push_back(file);
+//       }
+//       if (cleanup_fn != /*CreateNewBackup*/ 4) {
+//         // This exists after CreateNewBackup because it's deleted then
+//         // re-created.
+//         tmp_files_and_dirs.push_back(backupdir_ + "/" + next_private);
+//       }
+//
+//       OpenDBAndBackupEngine(false /* destroy_old_data */, false /* dummy */,
+//                             shared_option);
+//       // Need to call one of these explicitly to delete tmp files
+//       switch (cleanup_fn) {
+//         case 1:
+//           ASSERT_OK(backup_engine_->GarbageCollect());
+//           break;
+//         case 2:
+//           ASSERT_OK(backup_engine_->DeleteBackup(oldest_id));
+//           break;
+//         case 3:
+//           ASSERT_OK(backup_engine_->PurgeOldBackups(1));
+//           break;
+//         case 4:
+//           // Does a garbage collect if it sees that next private dir exists
+//           ASSERT_OK(backup_engine_->CreateNewBackup(db_.get()));
+//           break;
+//         default:
+//           assert(false);
+//       }
+//       CloseDBAndBackupEngine();
+//       for (std::string file_or_dir : tmp_files_and_dirs) {
+//         if (file_manager_->FileExists(file_or_dir) != Status::NotFound()) {
+//           FAIL() << file_or_dir << " was expected to be deleted." << cleanup_fn;
+//         }
+//       }
+//     }
+//   }
+// }
 
 TEST_F(BackupableDBTest, KeepLogFiles) {
   backupable_options_->backup_log_files = false;